doc/requirements.txt
# This file lists all the dependencies required to build the documentation.
# To update this file, follow these basic rules:
#   - Keep it sorted alphabetically
#   - Mark which docsets are using each dependency
#   - Only specify version information if strictly required
#   - Keep nrf/installation/recommended_versions.rst up to date

# Extension                       | NCS | Kconfig | Matter | MCUboot | nrfxlib | TF-M | Zephyr |
anytree                         # |     |         |        |         |         |      |   X    |
azure-storage-blob              # |  X  |         |        |         |         |      |        |
doxmlparser                     # |     |         |        |         |         |      |   X    |
m2r2                            # |     |         |        |         |         |      |        |
PyYAML                          # |  X  |         |        |         |         |      |   X    |
pykwalify                       # |     |         |        |         |         |      |   X    |
pytest                          # |     |         |        |         |         |      |   X    |
recommonmark                    # |     |         |   X    |    X    |         |      |        |
snowballstemmer<3.0.0           # https://github.com/snowballstem/snowball/issues/229
sphinx~=8.1                     # |  X  |    X    |   X    |    X    |    X    |  X   |   X    |
sphinx-autobuild                # |  X  |    X    |   X    |    X    |    X    |  X   |   X    |
sphinx-copybutton               # |  X  |         |        |         |         |      |   X    |
sphinx-ncs-theme<1.1            # |  X  |         |        |         |         |      |        |
sphinx-notfound-page>=1.0.0     # |  X  |         |        |         |         |      |   X    |
sphinx-tabs>=3.4                # |  X  |         |        |         |         |      |   X    |
sphinx-togglebutton             # |  X  |         |        |         |         |      |   X    |
sphinx_markdown_tables          # |     |         |   X    |         |         |      |        |
sphinxcontrib-mscgen            # |  X  |         |        |         |    X    |      |        |
sphinxcontrib-plantuml>=0.27    # |     |         |        |         |         |  X   |        |
west>=1.0.0                     # |     |         |        |         |         |      |   X    |
